+++ src/hooks/useConditional.js
@@ -11,7 +11,8 @@ export const useConditional = ({
   evaluate,
   evaluateWhen = [],
   dependsOn = [],
-  native = false
+  native = false,
+  evaluateOnMount = true
 }) => {
   // Grab the form controller
   const formController = useFormController();
@@ -25,7 +26,7 @@ export const useConditional = ({
 
   // Conditional state
   const [props, setProps] = useState(() => {
-    if (evaluate) {
+    if (evaluate && evaluateOnMount) {
       return evaluate({
         formState: formController.getFormState(),
         formApi: formController.getFormApi(),
@@ -57,14 +58,21 @@ export const useConditional = ({
   useFieldSubscription(
     event,
     fields,
-    target => {
-      logger(`re-evaluating conditional for ${name} because of ${target}`);
+    (target, triggers) => {
+      logger(
+        `re-evaluating ${name} because ${target} changed, triggerd by ${JSON.stringify(
+          triggers,
+          null,
+          2
+        )}`
+      );
       const res = evaluate({
         formState: formController.getFormState(),
         formApi: formController.getFormApi(),
         scope: scopeRef.current,
         dependsOn,
-        target
+        target,
+        triggers
       });
       setProps(res);
     },
@@ -75,8 +83,8 @@ export const useConditional = ({
 
   useEffect(
     () => {
-      if (evaluate) {
-        // When name changes we always check if relevant
+      if (evaluate && evaluateOnMount) {
+        // When name changes we always re evaluate
         setProps(
           evaluate({
             formState: formController.getFormState(),
@@ -93,7 +101,7 @@ export const useConditional = ({
   // Trigger evaluate on a reset of form
   useEffect(() => {
     const listener = () => {
-      if (evaluate) {
+      if (evaluate && evaluateOnMount) {
         setProps(
           evaluate({
             formState: formController.getFormState(),