+++ vitedocs/Pages/ApiReference/Arrays/ArrrayFieldSwap/ArrayFieldSwap.jsx
@@ -1,0 +1,26 @@
+import { useEffect } from 'react';
+import Example from './Example';
+import exampleCode from './Example.jsx?raw';
+import Code from '../../../../YourComponents/Code';
+import { SideBySide } from '../../../../SideBySide';
+import { Info } from '../../../../Info';
+
+export default function ArrayFieldComponent({ showItem = true }) {
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
+  return (
+    <>
+      <h1>
+        <code>ArrayField Swap</code>
+      </h1>
+      <Info>Sometimes you need to swap valuse around in an array.</Info>
+      <SideBySide
+        leftHeader={<h3>Example:</h3>}
+        rightHeader={<h3>Code:</h3>}
+        left={<Example />}
+        right={<Code links input1={exampleCode} />}
+      />
+    </>
+  );
+}