+++ vitedocs/Pages/ApiReference/GlobalForms/FormState/GlobalFormState.jsx
@@ -1,0 +1,32 @@
+import { useEffect } from 'react';
+import Example from './Example';
+import exampleCode from './Example.jsx?raw';
+import Code from '../../../../YourComponents/Code';
+import { SideBySide } from '../../../../SideBySide';
+import { Info } from '../../../../Info';
+
+export default function GlobalFieldState() {
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
+  return (
+    <>
+      <h1>
+        <code>Global Field State</code>
+      </h1>
+      <Info>
+        Sometimes you need to share state between forms. Below is an example
+        where you can name your forms and reference them while in the context of
+        Informed.
+      </Info>
+      <SideBySide
+        leftHeader={<h3>Example:</h3>}
+        rightHeader={<h3>Code:</h3>}
+        left={<Example />}
+        right={<Code links input1={exampleCode} />}
+      />
+      <br />
+      <br />
+    </>
+  );
+}